add new exposition crate
Adds the start of an exposition crate which provides a type for
holding a snapshot of the current metric values and a type that
can produce snapshots which may optionally be filtered down to a
subset of the registered metrics.

This gives us a single place to have a serializable snapshot type
and in the future can be extended to include additional exposition
related functionality.