exposition: make helper functions pub visible (#45)
The helper functions that allow accessing fields within the
snapshot regardless of variant are extremely useful in other
projects, particularly if you want to consume snapshots of either
flavor without caring about which flavor.

This change makes these functions publically visible so they do not
need to be implemented in other projects.